Leaderboards
Player will now be prompted to enter a name. Finish line trigger is in the scene, and will call to populate leaderboard list as well as take in the time it took to get there. Time starts when name is set. name is only set once and then never again. might have to figure out what happens if someone closes the game and then wants to play later, but for now, this is fine.